Version 0.3.0 Release (#18)
This release contains significant improvements. The old method of overriding method_added would have slowed down loading of large classes, since it needs to be run on every method that is defined on the class that includes Asynchronize. This version uses Module#prepend in order to override the methods.

This does have some trade offs, as we can no longer support Ruby < 2.3. I've changed the required ruby version in the Gem, so users below this version should not receive it.